requirements.txt
discord.py
requests
apscheduler
asyncio
